requirements.txt
numpy==1.23.5
pandas==1.5.2
joblib==1.1.1
PyYAML==6.0.2
scikit-learn==1.3.2
matplotlib==3.6.2
category_encoders==2.6.4
openpyxl==3.1.5
